explode entry points inside ./lib directory add doc block for lisibility , UPDATE README.md
