+++ package.json
@@ -1,7 +1,7 @@
 {
   "homepage": "https://https://github.com/Orivoir/express-live-reloading",
   "name": "express-live-reloading",
-  "version": "0.6.14",
+  "version": "0.7.0",
   "description": "Live reloading middleware express",
   "main": "express-live-reloading.js",
   "scripts": {},
